% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate a report on Top 5 stocks in current directory}
\usage{
generate_report(report_year, report_quarter, render = TRUE)
}
\arguments{
\item{report_year}{Year of report}

\item{report_quarter}{Quarter of report}
}
\description{
Fillin
}
