% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from Yahoo finance}
\usage{
get_data(report_year, report_quarter)
}
\arguments{
\item{report_year}{A numeric, the year period of stock performance}

\item{report_quarter}{A numeric, the quarter period in \code{report_year} of
stock performance}
}
\value{
A tibble, containing the historical data for stock performance
}
\description{
This function returns the historical end-of-day stock prices for the top
five stocks for the quarter indicated
}
