% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data into working space}
\usage{
load_data(report_year, report_quarter)
}
\arguments{
\item{report_year}{Reporting year}

\item{report_quarter}{Reporting quarter}
}
\description{
Load data into working space
}
